+++ packages/terra-list/src/SingleSelectList.jsx
@@ -17,6 +17,10 @@ const propTypes = {
    */
   hasChevrons: PropTypes.bool,
   /**
+   * Whether or not the child items should be displayed in sections with header and list of elements.
+   */
+  hasSections: PropTypes.bool,
+  /**
    * A callback event that will be triggered when selection state changes.
    */
   onChange: PropTypes.func,
@@ -26,6 +30,7 @@ const defaultProps = {
   children: [],
   isDivided: false,
   hasChevrons: false,
+  hasSections: false,
   onChange: undefined,
 };
 
@@ -33,13 +38,31 @@ class SingleSelectList extends React.Component {
   constructor(props) {
     super(props);
     this.handleOnChange = this.handleOnChange.bind(this);
-    this.state = { selectedIndex: SelectableList.Utils.initialSingleSelectedIndex(this.props.children) };
+    // state gets updated when getDerivedStateFromProps gets called right before each render.
+    this.state = {
+      selectedIndex: props.hasSections ?
+        SelectableList.Utils.initialSingleSelectedIndexWithSections(props.children) :
+        SelectableList.Utils.initialSingleSelectedIndex(props.children),
+      hasSections: false,
+    };
+  }
+
+  static getDerivedStateFromProps(props, state) {
+    if (props.hasSections !== state.hasSections) {
+      return {
+        hasSections: props.hasSections,
+      };
+    }
+    return null;
   }
 
   handleOnChange(event, index) {
     if (SelectableList.Utils.shouldHandleSingleSelect(this.state.selectedIndex, index)) {
       event.preventDefault();
-      this.setState({ selectedIndex: index });
+      this.setState(prevState => ({
+        ...prevState,
+        selectedIndex: index,
+      }));
       if (this.props.onChange) {
         this.props.onChange(event, index);
       }
@@ -48,15 +71,21 @@ class SingleSelectList extends React.Component {
 
   render() {
     const {
-      children, isDivided, onChange, hasChevrons, ...customProps
+      children,
+      isDivided,
+      onChange,
+      hasChevrons,
+      hasSections,
+      ...customProps
     } = this.props;
     return (
       <SelectableList
         {...customProps}
         isDivided={isDivided}
         onChange={this.handleOnChange}
-        selectedIndexes={[this.state.selectedIndex]}
         hasChevrons={hasChevrons}
+        hasSections={this.state.hasSections}
+        selectedIndexes={[this.state.selectedIndex]}
         disableUnselectedItems={false}
       >
         {children}
@@ -68,5 +97,6 @@ class SingleSelectList extends React.Component {
 SingleSelectList.propTypes = propTypes;
 SingleSelectList.defaultProps = defaultProps;
 SingleSelectList.Item = SelectableList.Item;
+SingleSelectList.Section = SelectableList.Section;
 
 export default SingleSelectList;