+++ packages/terra-list/tests/jest/MultiSelectList.test.jsx
@@ -50,6 +50,37 @@ it('should select an item when maxSelectionCount is 2', () => {
   expect(multiSelect).toMatchSnapshot();
 });
 
+it('should select an item across multiple sections when maxSelectionCount is 2', () => {
+  const sectionItems1 = [item1, item3];
+  const sectionItems2 = [item2, item4, item5];
+  const multiSelectList = (
+    <MultiSelectList hasSections maxSelectionCount={2}>
+      <MultiSelectList.Section
+        headerContent="Header1"
+        listItems={sectionItems1}
+      />
+      <MultiSelectList.Section
+        headerContent="Header2"
+        listItems={sectionItems2}
+      />
+    </MultiSelectList>
+  );
+
+  const multiSelect = mount(multiSelectList);
+
+  multiSelect.find('.stuff1').first().simulate('click', { preventDefault() {} });
+  expect(multiSelect).toMatchSnapshot();
+
+  multiSelect.find('.stuff2').first().simulate('click', { preventDefault() {} });
+  expect(multiSelect).toMatchSnapshot();
+
+  multiSelect.find('.stuff3').first().simulate('click', { preventDefault() {} });
+  expect(multiSelect).toMatchSnapshot();
+
+  multiSelect.find('.stuff4').first().simulate('click', { preventDefault() {} });
+  expect(multiSelect).toMatchSnapshot();
+});
+
 it('should mount with no items', () => {
   const multiSelect = mount(<MultiSelectList />);
   expect(multiSelect).toMatchSnapshot();
@@ -57,5 +88,18 @@ it('should mount with no items', () => {
 
 it('should mount with one item', () => {
   const multiSelect = mount(<MultiSelectList>{[item1]}</MultiSelectList>);
+  expect(multiSelect).toMatchSnapshot();
+});
+
+it('should mount with no items with section headers', () => {
+  const multiSelect =
+    mount(<MultiSelectList hasSections><MultiSelectList.Section headerContent="Header1" /></MultiSelectList>);
+  expect(multiSelect).toMatchSnapshot();
+});
+
+
+it('should mount with one items with Section Headers', () => {
+  const multiSelect =
+    mount(<MultiSelectList hasSections><MultiSelectList.Section headerContent="Header1" listItems={[item1]} /></MultiSelectList>);
   expect(multiSelect).toMatchSnapshot();
 });